requirements.txt
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
asyncpg==0.30.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
codex_cli==1.0.1
colorama==0.4.6
coverage==7.10.6
duckdb==1.3.2
et_xmlfile==2.0.0
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.3.0
playwright==1.55.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.8.29
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
